Move bootstrapper back to /js
