+++ webapp/src/js/bootstrapper/translator.js
